chore: Use Core phase v2
* Card ID: CCT-1321

While the RPM-shipped Core behaves mostly the same as the CDN Core,
there are behavioral differences between the two.

Phase 2 drops the 'update' phase, where an egg is fetched and
GPG-verified. In the end of 'collect_and_upload' phase, v2 drops
egg rotation code, because there are no eggs to rotate.